+++ src/common/utils/index.js
@@ -1,3 +1,4 @@
+export * from './fetchLocations';
 export * from './getDehydratedPiece';
 export * from './getHydratedPieces';
 export * from './ifMissingPermanentLoanTypeId';